+++ src/main/java/org/logstash/beats/Runner.java
@@ -1,17 +1,13 @@
 package org.logstash.beats;
 
-import io.netty.util.ResourceLeakDetector;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.logstash.netty.PrivateKeyConverter;
+import org.apache.log4j.Logger;
 import org.logstash.netty.SslSimpleBuilder;
-
-import java.io.FileInputStream;
 
 
 public class Runner {
     private static final int DEFAULT_PORT = 5044;
-    public static Logger logger = LogManager.getLogger(Runner.class.getName());
+    private final static Logger logger = Logger.getLogger(Runner.class);
+
 
 
     static public void main(String[] args) throws Exception {
@@ -22,7 +18,7 @@ public class Runner {
 
         Server server = new Server(DEFAULT_PORT);
 
-        if(args.length > 0 && args[0].equals("ssl")) {
+            if(args.length > 0 && args[0].equals("ssl")) {
             logger.debug("Using SSL");
 
             String sslCertificate = "/Users/ph/es/certificates/certificate.crt";
@@ -30,13 +26,9 @@ public class Runner {
             String noPkcs7SslKey = "/Users/ph/es/certificates/certificate.key";
             String[] certificateAuthorities = new String[] { "/Users/ph/es/certificates/certificate.crt" };
 
-            PrivateKeyConverter converter = new PrivateKeyConverter(noPkcs7SslKey, null);
 
-            logger.debug("SSLCertificate: " + sslCertificate);
-            logger.debug("SSLKey: " + sslKey);
 
-//            SslSimpleBuilder sslBuilder = new SslSimpleBuilder(sslCertificate, sslKey, null)
-            SslSimpleBuilder sslBuilder = new SslSimpleBuilder(new FileInputStream(sslCertificate), converter.convert(), null)
+            SslSimpleBuilder sslBuilder = new SslSimpleBuilder(sslCertificate, sslKey, null)
                     .setProtocols(new String[] { "TLSv1.2" })
                     .setCertificateAuthorities(certificateAuthorities)
                     .setHandshakeTimeoutMilliseconds(10000);